fixes and blood balance

+++ Dragon's Trawel/src/Item.java
@@ -9,7 +9,7 @@ public abstract class Item implements java.io.Serializable{
 	
 	protected int level;
 	public int bloodSeed = extra.randRange(0,2000);
-	private int bloodCount = 0;
+	private float bloodCount = 0;
 
 	//abstracted instance methods
 	/**
@@ -72,10 +72,10 @@ public abstract class Item implements java.io.Serializable{
 	}
 	
 	public int getBloodCount() {
-		return bloodCount;
+		return (int)bloodCount;
 	}
 	
-	public void addBlood(int i) {
+	public void addBlood(float i) {
 		bloodCount+=i;
 		if (bloodCount > 16) {
 			bloodCount = 16;
